ENH: Add SetOutputScalarType() method to vtkImageReslice.
This is an experimental modification to vtkImageReslice.  The
motivation behind it is that interpolation is always done with
floating-point, so casting the interpolated result back to the
input type is lossy.  By producing an output of type Float or
Double for use by vtkImageMapToColors, the banding artifact
can be eliminated.